docsage/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
qdrant-client
llama-index
llama-index-core
llama-index-embeddings-openai
llama-index-readers-file
llama-index-vector-stores-qdrant
openai
sentence-transformers
rank-bm25
ragas
datasets
ipykernel
jinja2
httpx
orjson
